MDEV-17515: GTID Replication in optimistic mode deadlock
Problem:
=======
In slave_parallel_mode=optimistic configuration, when admin commands and
DML operation on the same table are scheduled simultaneously for execution,
it results in lock conflict and slave server either hangs due to deadlock
or goes down with an assert.

Analysis:
========
Admin commands acquire table locks and these locks are not detected
by parallel replication deadlock detection-and-handling mechanism.
Hence they must not be allowed to execute in parallel with DMLs.

Fix:
===
Mark admin commands as DDL as DDLs are not scheduled for parallel
execution.